% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adae}
\alias{adae}
\title{Adverse Events Analysis Data}
\format{
A data frame with 756 rows and 55 variables:
}
\source{
\url{https://github.com/phuse-org/phuse-scripts/raw/master/data/adam/cdisc-split}
}
\usage{
adae
}
\description{
A dataset containing the adverse event information of 148 unique subjects
}
\keyword{datasets}
