% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adlbh}
\alias{adlbh}
\title{Hematology Laboratory Analysis Data}
\format{
A data frame with 34766 rows and 46 variables:
}
\source{
\url{https://github.com/phuse-org/phuse-scripts/raw/master/data/adam/cdisc-split}
}
\usage{
adlbh
}
\description{
A dataset containing the laboratory information of 170 unique subjects
}
\keyword{datasets}
