% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adsl}
\alias{adsl}
\title{Subject Level Analysis Data}
\format{
A data frame with 170 rows and 50 variables:
}
\source{
\url{https://github.com/phuse-org/phuse-scripts/raw/master/data/adam/cdisc-split}
}
\usage{
adsl
}
\description{
A dataset containing the subject level information of 170 unique subjects
}
\keyword{datasets}
