% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adtte}
\alias{adtte}
\title{Time to Event Analysis Data}
\format{
A data frame with 170 rows and 26 variables:
}
\source{
\url{https://github.com/phuse-org/phuse-scripts/raw/master/data/adam/cdisc-split}
}
\usage{
adtte
}
\description{
A dataset containing the time to event information of 170 unique subjects
}
\keyword{datasets}
