% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ae_rate_plot.R
\name{ae_rate_plot}
\alias{ae_rate_plot}
\title{A Rate plot comparing Adverse Events (preferred terms) of interest by treatment}
\usage{
ae_rate_plot(adae, adsl, reord_type_rateplot, title, subtitle, caption)
}
\arguments{
\item{adae}{Adverse Event dataset}

\item{adsl}{Subject Level dataset}

\item{reord_type_rateplot}{A string shows how to reorder the `AEDECOD` on the plot.
Options are `diff` (arrange based on difference between treatment and placebo),
`trt` (arranged based on the treatment value), and
`placebo` (arranged based on the Placebo value).}

\item{title}{A string shows the main title of plot.}

\item{subtitle}{A string shows the subtitle of plot.}

\item{caption}{A string shows the caption of plot.}
}
\value{
A ggplot object
}
\description{
A Rate plot comparing Adverse Events (preferred terms) of interest by treatment
}
\examples{
\dontrun{
 ae_rate_plot(adae = adae,
                adsl = adsl)
}
}
\keyword{internal}
