% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_ae_summary.R
\name{ae_summary}
\alias{ae_summary}
\title{Comparing Adverse Events (preferred terms) of interest by treatment}
\usage{
ae_summary(
  adae,
  adsl,
  kind = "forest",
  pval_cutoff = 1,
  adjpval_cutoff = 0.9,
  reord_type_rateplot = "diff",
  title = "",
  subtitle = "",
  caption = ""
)
}
\arguments{
\item{adae}{Adverse Event dataset}

\item{adsl}{Subject Level dataset}

\item{kind}{The type of visualization. Options are `forest`, `rate`,
`volcano_pval`, and `volcano_adjpval`. Default is `forest`.}

\item{pval_cutoff}{A cut-off value for filtering the hazard output results
on `p.value < pval_cutoff`. No effects on the `rate` and `swimmer` plots.
Default is `1`.}

\item{adjpval_cutoff}{A cut-off value for showing the text on the `volcano_pval` and `volcano_adjpval outputs`.
It will apply a filtration on `adjpvalue < adjpval_cutoff`.
No effects on the `rate` and `swimmer` plots. Default is `0.9`.}

\item{reord_type_rateplot}{}

\item{title}{A string shows the main title plot.}

\item{subtitle}{A string shows the subtitle plot.}

\item{caption}{A string shows the caption of plot.}
}
\value{
A ggplot object
}
\description{
This function compares different adverse events using multiple visualization
methods: Forest plot, Rate plot, Swimmer plot, Volcano Plot for P-values, and Volcano Plots for Adjusted P-values.
}
\note{
for the `forest`, `volcano_pval`, and `volcano_adjpval`, this function
calculates hazard ratio based on the following `Cox` regression model:
\code{coxph(Surv(AVAL, CNSR) ~ TRT01A)}
}
\examples{
\dontrun{

 load("data/adsl.rda")
 load("data/adae.rda")

 ae_summary(adae = adae,
            adsl = adsl,
            kind = "forest",
            pval_cutoff = 0.6
           )

 ae_summary(adae = adae,
            adsl = adsl,
            kind = "rate"
           )

 ae_summary(adae = adae,
            adsl = adsl,
            kind = "swimmer"
           )

 ae_summary(adae = adae,
            adsl = adsl,
            kind = "volcano_pval",
            pval_cutoff = 1,
            adjpval_cutoff = 0.9
           )


 ae_summary(adae = adae,
            adsl = adsl,
            kind = "volcano_adjpval",
            pval_cutoff = 1,
            adjpval_cutoff = 0.9
            )
}
}
