% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ae_swimmer_plot.R
\name{ae_swimmer_plot}
\alias{ae_swimmer_plot}
\title{A swimmer plot comparing Adverse Events (preferred terms) of interest by treatment}
\usage{
ae_swimmer_plot(adae, adsl, title, subtitle, caption)
}
\arguments{
\item{adae}{Adverse Event dataset}

\item{adsl}{Subject Level dataset}

\item{title}{A string shows the main title plot.}

\item{subtitle}{A string shows the subtitle plot.}

\item{caption}{A string shows the caption of plot.}
}
\value{
A ggplot object
}
\description{
A swimmer plot comparing Adverse Events (preferred terms) of interest by treatment
}
\examples{
\dontrun{
 ae_rate_plot(adae = adae,
                adsl = adsl)
}
}
\keyword{internal}
