% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_lab_spaghetti.R
\name{lab_spaghetti}
\alias{lab_spaghetti}
\title{A laboratory spaghetti plot}
\usage{
lab_spaghetti(
  adlb,
  param_filter,
  axis_tick_size = 8,
  x_axis_wrap = 6,
  selected_visit = c("Baseline", "Week 2", "Week 4", "Week 6", "Week 8",
    "End of Treatment")
)
}
\arguments{
\item{adlb}{Laboratory dataset}

\item{param_filter}{Character. The filtration applies on `PARAM` variable.}

\item{axis_tick_size}{the size of the axis tick.}

\item{x_axis_wrap}{wrap long text in the x axis.}

\item{selected_visit}{Vector of characters. The filtration on `AVISIT` variable.}
}
\value{
A ggplot object
}
\description{
This functions creates a spaghetti plot for each lab parameter
}
\examples{
\dontrun{


}
}
