% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_lab_summary.R
\name{lab_summary}
\alias{lab_summary}
\title{A laboratory level summary bar plot}
\usage{
lab_summary(
  adlb,
  kind = "abnormal",
  param_filter,
  title = NULL,
  shift_breaks_vec = NULL,
  selected_visit = c("Baseline", "Week 2", "Week 4", "Week 6", "Week 8",
    "End of Treatment")
)
}
\arguments{
\item{adlb}{Laboratory dataset}

\item{kind}{The type of visualization. Options are `abnormal`, `box`, and `shift`. Default is `abnormal`.}

\item{param_filter}{Character. The filtration applies on `PARAM` variable.}

\item{title}{the main title of the plot.}

\item{shift_breaks_vec}{the break x & y axis for shift plot.}

\item{selected_visit}{Vector of characters. The filtration on `AVISIT` variable.}
}
\value{
A ggplot object
}
\description{
This functions creates a barplot for each treatment that shows
the percentage subjects below the lower limit of normal (LLN) and above
the upper limit of normal (ULN)
}
\examples{
\dontrun{

 load("data/adlbc.rda")

 lab_summary(adlb = adlbc,
             kind = "abnormal",
             param_filter = "Cholesterol (mmol/L)",
             selected_visit = c("Baseline", "Week 2", "Week 4", "Week 6", "Week 8", "End of Treatment")
            )

 lab_summary(adlb = adlbc,
             kind = "box",
             param_filter = "Cholesterol (mmol/L)",
             selected_visit = c("Baseline", "Week 2", "Week 4", "Week 6", "Week 8", "End of Treatment")
            )

}
}
