% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_patient_profile.R
\name{patient_profile}
\alias{patient_profile}
\title{A Patient Profile function}
\usage{
patient_profile(
  data,
  usubjid = "02-718-1371",
  kind = "summary",
  ncol_lab = 1,
  filter_safety_flag_lab = "Y",
  transform_lab = "log",
  height_ae = 250,
  order_servere_ae = c("MILD", "MODERATE", "SEVERE"),
  color_ae = sample_colors[c("Blue Light", "Apricot", "Red")],
  color_cm = sample_colors[c("Purple")],
  str_wrap_ae_cm = 25,
  params_lab = c("Alanine Aminotransferase (U/L)", "Aspartate Aminotransferase (U/L)",
    "Bilirubin (umol/L)"),
  selected_visit = c("Baseline", "Week 2", "Week 4", "Week 6", "Week 8",
    "End of Treatment")
)
}
\arguments{
\item{data}{dataset}

\item{kind}{The type of visualization. Options are `summary`, `lab`, `ae`, `cm`. The data domain should be matched with the kind of visualization.}
}
\value{
A ggplot object
}
\description{
This functions creates a spaghetti plot for each lab parameter
}
\examples{
\dontrun{


}

}
